Create README.md
****Explanation of Outcome****
-From the dataset analysed above,we found the following insights;
	*we find that there is only few people who smoke from the regions recorded.
	*Also,when comparing people who smoke with respect to sex,we find that most people who smoke are men compared to female.
	*When ploting a histogram for the age that makes up the population,we fing age goup in the population ranges from 10-65 years
	*We also find that,most smokers are from the regionof southeast while southwest and Northwest people have fewer smokers recorded.
	*Also from the population data, its found that, most people smokers are adults.

  
**PROCESS OF PREDICTION**
Splits the data into features (stored in X) and target (stored in y).
Splits the data into training and testing sets (80% for training and 20% for testing).
Fits a linear regression model to the training data using the LinearRegression class from scikit-learn.
Makes predictions on the test data using the predict method.
Calculates the mean squared error (MSE) between the actual target values and the predicted values. The MSE is a measure of the quality of the model's predictions.

The mean squared error (MSE) is a common metric for evaluating the performance of a regression model. It measures the average of the squared differences between the actual target values and the predicted values. A lower MSE indicates a better-performing model.

The MSE is calculated by taking the difference between the actual and predicted values, squaring them, and then taking the mean of the squared differences. This gives a measure of the average magnitude of the error in the model's predictions, regardless of the direction of the error. The MSE penalizes large errors more than small errors